Beginnings of database implementation
Added database methods that correspond to each protocol action in IWDB.ts. Added an express route that handles creating a room in the databasein protoserver.ts, as well as a basic socket generator and roomcode generator just to trial things. Modified Host.tsx to request creation of a room, the room request probably doesn't need it's own function. Need to set room information with state then use that to create the websocket connection.