Finished host room creation for now.
Host now sends a request to protoserver.ts on the '/create' endpoint, which generates a new room entry in the database, and then sends that room to the Host. Once fetched, it then sets the room on the host and then sets the websocket based on the room information. Participant will need to apply a different endpoint, '/join', which will be created next.