Applied random string room code generation and fixed funny bug that made my program crash.
I made some changes to CreateNewCode, so it performs a search on the DB before providing the code. Did this with a do while loop but because it wasn't awaiting the RoomSearch it just kept appending characters to the 'output string' and  made the program run out of heap size and crash :( . So i fixed that up and made it an async function. No more crashes :)